+++ packages/adapter-webpack/.eslintrc.js
@@ -1,0 +1,9 @@
+module.exports = {
+  root: false,
+  parserOptions: {
+    parser: 'babel-eslint'
+  },
+  extends: [
+    'plugin:vue/recommended'
+  ]
+}