+++ packages/adapter-webpack/test/fixtures/react/server-render.js
@@ -1,0 +1,5 @@
+const { renderToString } = require('react-dom/server')
+
+module.exports = function serverRender (Component, props) {
+  return renderToString(Component(props))
+}