+++ packages/adapter-webpack/test/fixtures/react/template.jsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Template = props => {
+  return (
+    <p>
+      {props.myData}
+    </p>
+  )
+}
+
+export default Template