+++ packages/adapter-webpack/test/fixtures/vue/server-render.js
@@ -1,0 +1,14 @@
+const Vue = require('vue')
+const { createRenderer } = require('vue-server-renderer')
+
+const renderer = createRenderer()
+
+module.exports = function serverRender (Component, props) {
+  return renderer.renderToString(
+    new Vue({
+      render (h) {
+        return h(Component, { props })
+      }
+    })
+  )
+}