+++ test/project/src/elements/label/component.config.js
@@ -72,15 +72,9 @@ module.exports = {
       }
     },
     {
-      file: 'label-vue.js',
-      title: 'Label (Vue JS)',
+      file: 'label.vue',
+      title: 'Label (Vue)',
       label: 'A1.6',
-      tags: ['Vue']
-    },
-    {
-      file: 'label-vue-sfc.vhtml',
-      title: 'Label (Vue SFC)',
-      label: 'A1.7',
       tags: ['Vue']
     }
   ]