+++ test/project/src/elements/label/label-register.vue.js
@@ -1,8 +1,0 @@
-import Vue from 'vue'
-import props from './props'
-
-Vue.component('my-label', {
-  props,
-
-  template: '<label class="label" :for="id">{{ title }}</label>'
-})